requirements.txt
coverage==4.5.1
codecov==2.0.15
django==1.11.18
django-pipeline==1.6.14
django-widget-tweaks==1.4.2
django-dotenv==1.4.2
django-debug-toolbar==1.9.1
mysqlclient==1.3.13
django-redis==4.9.0
pytz==2018.5
python-dateutil==2.7.3
mock==2.0.0
djangorestframework==3.8.2
phonenumbers==8.9.9
twilio==6.14.9
schedule==0.5.0
pyngrok==1.3.1